@@ -1,6 +1,6 @@
 {
   "name": "acikkaynak-website",
-  "version": "0.9.0",
+  "version": "0.10.0",
   "description": "",
   "keywords": [
     "typescript",
@@ -15,10 +15,14 @@
     "url": "https://github.com/acikkaynak/acikkaynak-website/issues",
     "email": "eser@ozvataf.com"
   },
+  "main": "dist/index.js",
   "scripts": {
+    "init": "cp .env.default .env",
     "start": "cross-env NODE_ENV=development node ./server.js",
+    "lint": "eslint ./src/",
+    "test": "jest",
     "bundle": "webpack",
-    "bundle:prod": "cross-env NODE_ENV=production webpack --config ./webpack.config.prod.js"
+    "bundle:prod": "webpack -p"
   },
   "license": "Apache-2.0",
   "contributors": [
@@ -34,49 +38,53 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@types/react": "^16.0.2",
-    "autoprefixer": "^7.1.2",
-    "bootstrap-sass": "^3.3.7",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^5.1.1",
-    "css-loader": "^0.28.4",
+    "@types/jest": "^23.1.0",
+    "@types/react": "^16.3.17",
+    "autoprefixer": "^8.6.2",
+    "bulma": "^0.7.1",
+    "clean-webpack-plugin": "^0.1.19",
+    "copy-webpack-plugin": "^4.5.1",
+    "cross-env": "^5.2.0",
+    "css-loader": "^0.28.11",
     "cssnano": "^3.10.0",
-    "es6-cachemanager": "^0.0.2",
-    "es6-promise": "^4.0.5",
-    "eslint": "^4.3.0",
-    "eslint-config-eser": "^2.0.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "express": "^4.15.3",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
+    "es6-cachemanager": "0.0.2",
+    "es6-promise": "^4.2.4",
+    "eslint": "^4.19.1",
+    "eslint-config-eser": "^2.0.7",
+    "eslint-config-eser-react": "^2.0.7",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.9.1",
+    "express": "^4.16.3",
+    "file-loader": "^1.1.11",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.2.1",
+    "jest": "^23.1.0",
     "json-loader": "^0.5.7",
-    "node-sass": "^4.5.3",
+    "mini-css-extract-plugin": "^0.4.0",
+    "node-sass": "^4.9.0",
     "path-browser": "^2.2.1",
-    "postcss": "^6.0.9",
-    "postcss-loader": "^2.0.6",
+    "postcss": "^6.0.22",
+    "postcss-loader": "^2.1.5",
     "postcss-modules": "^1.1.0",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "prop-types": "^15.6.1",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "react-eventmanager": "0.0.3",
-    "react-intl": "^2.3.0",
-    "react-markdown": "^2.5.0",
-    "react-router-dom": "^4.1.2",
-    "sass-loader": "^6.0.6",
-    "servicemanager": "^0.0.3",
-    "style-loader": "^0.18.2",
-    "ts-loader": "^2.3.2",
-    "typescript": "^2.4.2",
-    "uglifyjs-webpack-plugin": "^1.1.1",
-    "webpack": "^3.5.3",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-middleware": "^1.12.0",
+    "react-intl": "^2.4.0",
+    "react-markdown": "^3.4.1",
+    "react-router-dom": "^4.3.1",
+    "sass-loader": "^7.0.3",
+    "servicemanager": "^0.1.0",
+    "style-loader": "^0.21.0",
+    "ts-jest": "^22.4.6",
+    "ts-loader": "^4.4.1",
+    "typescript": "^2.9.2",
+    "webpack": "^4.12.0",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-cli": "^3.0.6",
+    "webpack-dev-middleware": "^3.1.3",
     "webpack-dotenv-plugin": "^2.0.2",
-    "webpack-hot-middleware": "^2.18.2",
-    "whatwg-fetch": "^2.0.3"
+    "webpack-hot-middleware": "^2.22.2",
+    "whatwg-fetch": "^2.0.4"
   }
 }
